+++ networking/eth2/src/test/java/tech/pegasys/teku/networking/eth2/rpc/beaconchain/BeaconChainMethodsTest.java
@@ -93,9 +93,7 @@ public class BeaconChainMethodsTest {
     final BeaconChainMethods methods = getMethods(TestSpecFactory.createMinimalAltair());
 
     assertThat(methods.beaconBlocksByRange().getIds())
-        .containsExactly(
-            "/eth2/beacon_chain/req/beacon_blocks_by_range/2/ssz_snappy",
-            "/eth2/beacon_chain/req/beacon_blocks_by_range/1/ssz_snappy");
+        .containsExactly("/eth2/beacon_chain/req/beacon_blocks_by_range/2/ssz_snappy");
   }
 
   @Test
@@ -103,7 +101,7 @@ public class BeaconChainMethodsTest {
     final BeaconChainMethods methods = getMethods();
 
     assertThat(methods.beaconBlocksByRange().getIds())
-        .containsExactly("/eth2/beacon_chain/req/beacon_blocks_by_range/1/ssz_snappy");
+        .containsExactly("/eth2/beacon_chain/req/beacon_blocks_by_range/2/ssz_snappy");
   }
 
   @Test
@@ -111,17 +109,7 @@ public class BeaconChainMethodsTest {
     final BeaconChainMethods methods = getMethods(TestSpecFactory.createMinimalAltair());
 
     assertThat(methods.beaconBlocksByRoot().getIds())
-        .containsExactly(
-            "/eth2/beacon_chain/req/beacon_blocks_by_root/2/ssz_snappy",
-            "/eth2/beacon_chain/req/beacon_blocks_by_root/1/ssz_snappy");
-  }
-
-  @Test
-  public void shouldCreateUnversionedBlocksByRootMethodWithAltairDisabled() {
-    final BeaconChainMethods methods = getMethods();
-
-    assertThat(methods.beaconBlocksByRoot().getIds())
-        .containsExactly("/eth2/beacon_chain/req/beacon_blocks_by_root/1/ssz_snappy");
+        .containsExactly("/eth2/beacon_chain/req/beacon_blocks_by_root/2/ssz_snappy");
   }
 
   @Test